pyproject.toml
[tool.poetry]
name = "music-generator"
version = "0.1.0"
description = ""
authors = ["astranero <esa.taheresa@gmail.com>"]
readme = "README.md"
packages = [{include = "music_generator"}]

[tool.poetry.dependencies]
python = "^3.10"
mido = "^1.2.10"
midi2audio = "^0.1.1"
fluidsynth = "^0.2"
pygame = "^2.1.2"
pytest = "^7.2.1"
invoke = "^2.0.0"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
